Update menu
A lot has changed with the menu. Instead of using a struct as a menuitem we now have separate classes for each type that can be added to the menu. This was implemented in order to reduce the 'overhead' for managing the menuitems. Now on an event the menu lets each menuitem handle the event by calling the 	rigger() function for the menuitem.